feat: 完整社保养老金计算器APP
- 36个省市区社保数据（缴费基数上下限、计发基数、费率）
- 退休年龄自动计算（渐进式延迟退休 2025-2039）
- 按年龄输入快速模式（滑块+文本框）
- 养老金精算：基础养老金+个人账户+过渡性养老金
- 社保月缴费明细展示
- Flutter analyze: 0 errors, 0 warnings

+++ lib/models/models.dart
@@ -137,7 +137,7 @@ class PensionResult {
   final double totalMonthlyPension; // 合计月养老金
   final double provinceBase; // 当地计发基数
   final double indexedSalary; // 本人指数化月平均工资
-  final int finalYears; // 最终缴费年限
+  final double finalYears; // 最终缴费年限
   final double finalAccountBalance; // 退休时个人账户余额
 
   PensionResult({
